src/hooks/usePerformance.ts
import { useCallback, useRef } from 'react';
import { debounce, throttle } from '../services/performance/debounce';

export function usePerformance() {
  const analysisDebounced = useRef(
    debounce((text: string) => {
      // Implement debounced analysis
    }, 500)
  ).current;

  const transcriptionThrottled = useRef(
    throttle((text: string) => {
      // Implement throttled transcription
    }, 100)
  ).current;

  const optimizeImage = useCallback(async (file: File): Promise<File> => {
    // Implement image optimization
    return file;
  }, []);

  return {
    analysisDebounced,
    transcriptionThrottled,
    optimizeImage
  };
}